requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
urllib3>=2.0.0
pgvector>=0.2.1
pg8000>=1.29.0